+++ src/Http/Controllers/L5ebtapiController.php
@@ -14,30 +14,26 @@ class L5ebtapiController extends Controller
 {
 
     // Declare private variables
-    private $api_url = '';
+    private $api_url = 'https://api.ebay.com/ws/api.dll';
     private $api_verify_ssl = true;
-    private $api_compatibility_level = '';
+    private $api_compatibility_level = '971';
     private $api_error_language = 'US';
     private $api_warning_level = 'LOW';
+    private $api_runame = '';
     private $api_user_token = '';
     private $api_dev_id;
     private $api_app_id;
     private $api_cert_id;
-    private $api_site_id;
+    private $api_site_id = '0';
 
     /**
-     * Create a new controller instance.
+     * Create a new L5ebtapiController instance.
      *
      * @return void
      */
     public function __construct($attributes)
     {
-
-        if (isset($attributes['api_url']) && empty($attributes['api_url']) || is_null($attributes['api_url'])) {
-
-            $this->api_url = 'https://api.ebay.com/ws/api.dll';
-
-        } else {
+        if (isset($attributes['api_url']) && strlen($attributes['api_url']) > 0) {
 
             $this->api_url = $attributes['api_url'];
 
@@ -51,31 +47,28 @@ class L5ebtapiController extends Controller
             $this->api_verify_ssl = true;
 
         }
-        if (isset($attributes['api_compatibility_level']) && empty($attributes['api_compatibility_level']) || is_null($attributes['api_compatibility_level'])) {
-
-            $this->api_compatibility_level = '971';
-
-        } else {
+        if (isset($attributes['api_compatibility_level']) && strlen($attributes['api_compatibility_level']) > 0) {
 
             $this->api_compatibility_level = $attributes['api_compatibility_level'];
 
         }
-        if (isset($attributes['api_error_language']) && empty($attributes['api_error_language']) || is_null($attributes['api_error_language'])) {
-
-            $this->api_error_language = 'US';
-
-        } else {
+        if (isset($attributes['api_error_language']) && strlen($attributes['api_error_language']) > 0) {
 
             $this->api_error_language = $attributes['api_error_language'];
 
         }
-        if (isset($attributes['api_warning_level']) && empty($attributes['api_warning_level']) || is_null($attributes['api_warning_level'])) {
+        if (isset($attributes['api_warning_level']) && strlen($attributes['api_warning_level']) > 0) {
 
-            $this->api_warning_level = 'LOW';
+            $this->api_warning_level = $attributes['api_warning_level'];
+
+        }
+        if (isset($attributes['api_runame']) && empty($attributes['api_runame']) || is_null($attributes['api_runame'])) {
+
+            $this->api_runame = '';
 
         } else {
 
-            $this->api_warning_level = $attributes['api_warning_level'];
+            $this->api_runame = $attributes['api_runame'];
 
         }
         if (isset($attributes['api_user_token']) && empty($attributes['api_user_token']) || is_null($attributes['api_user_token'])) {
@@ -102,11 +95,7 @@ class L5ebtapiController extends Controller
             $this->api_cert_id = $attributes['api_cert_id'];
 
         }
-        if (isset($attributes['api_site_id']) && empty($attributes['api_site_id']) || is_null($attributes['api_site_id'])) {
-
-            $this->api_site_id = '0';
-
-        } else {
+        if (isset($attributes['api_site_id']) && strlen($attributes['api_site_id']) > 0) {
 
             $this->api_site_id = $attributes['api_site_id'];
 
@@ -114,9 +103,16 @@ class L5ebtapiController extends Controller
 
     }// END constructor
 
+    /**
+     * Method: getEbayOfficialTime() - get the eBay official time API call.
+     *
+     * @param $multiPartImageData the image data. Acceptable formats (jpg, gif, png)
+     * @param $image_name The name the uploaded image will have on the eBay Picture Services server.
+     * @return string The URL of the image on the eBay Picture Services server OR an error message string if an error
+     * occurs.
+     */
     public function getEbayOfficialTime()
     {
-        //$timestamp = '';
 
         $request_body = '<?xml version="1.0" encoding="utf-8"?>
                         <GeteBayOfficialTimeRequest xmlns="urn:ebay:apis:eBLBaseComponents">
@@ -129,9 +125,9 @@ class L5ebtapiController extends Controller
 
         if (stristr($responseXml, 'HTTP 404')) {
 
-            Log::error('eBay API Call: FetchToken: User: ' . Auth::user()->id . ' 404 Not Found');
+            Log::error('eBay API Call: getEbayOfficialTime() 404 Not Found');
 
-            $message = ['error' => '404 Not Found. Please verify all eBay API settings are correct and try again.'];
+            $message = ['Error: 404 Not Found. Please verify all eBay API settings are correct and try again.'];
 
             return $message;
 
@@ -139,10 +135,10 @@ class L5ebtapiController extends Controller
 
         if ($responseXml == '') {
 
-            Log::error('eBay API Call: FetchToken: User: ' . Auth::user()->id . ' Error sending request' .
-                'XML response is an empty string');
+            Log::error('eBay API Call: getEbayOfficialTime() Error sending request' .
+                'the XML response is an empty string');
 
-            $message = ['error' => 'Error sending request XML response is an empty string. Please verify all eBay' .
+            $message = ['Error: The XML response is an empty string. Please verify all eBay' .
                 'API settings are correct and try again.'];
 
             return $message;
@@ -189,15 +185,23 @@ class L5ebtapiController extends Controller
         {
             $xml = simplexml_load_string($responseDoc->saveXML());
 
-            if ($xml->Timestamp) {
+            if ($xml->Ack && ((string)$xml->Ack == 'Success')) {
 
-                $timestamp = (string) $xml->Timestamp;
+                if ($xml->Timestamp) {
 
-            }
-            else {
+                    $timestamp = (string)$xml->Timestamp;
 
-                $timestamp = 'An error occurred while processing the getEbayOfficialTime() request. Please verify all' .
-                    'eBay API settings are correct and try the request again.';
+                } else {
+
+                    $timestamp = 'Error: An error occurred While processing the getEbayOfficialTime() request. Please' .
+                        'verify all eBay API settings are correct and try the request again.';
+
+                }
+
+            } else {
+
+                $timestamp = 'Error: An error occurred While processing the getEbayOfficialTime() request. Please' .
+                    'verify all eBay API settings are correct and try the request again.';
 
             }
 
@@ -208,13 +212,96 @@ class L5ebtapiController extends Controller
     }// END getEbayOfficialTime()
 
     /**
-     * Make an eBay API request
+     * Method: uploadSiteHostedPictures($multiPartImageData, $image_name) - Upload an image to the eBay Picture Service.
      *
-     * @param $method the eBay API method
+     * @param $multiPartImageData the image data. Acceptable formats (jpg, gif, png)
+     * @param $image_name The name the uploaded image will have on the eBay Picture Services server.
+     * @return string The URL of the image on the eBay Picture Services server OR an error message string if an error
+     * occurs.
+     */
+    public function uploadSiteHostedPictures($multiPartImageData, $image_name)
+    {
+
+        ///Build the request XML request which is first part of multi-part POST
+        $xmlReq = '<?xml version="1.0" encoding="utf-8"?>
+                           <UploadSiteHostedPicturesRequest xmlns="urn:ebay:apis:eBLBaseComponents">
+                           <Version>' . $this->api_compatibility_level . '</Version>
+                           <RequesterCredentials>
+                           <eBayAuthToken>' . $this->api_user_token . '</eBayAuthToken>
+                           </RequesterCredentials>
+                           <PictureName>' . $image_name . '</PictureName>
+                           <PictureSet>Standard</PictureSet>
+                           <PictureUploadPolicy>ClearAndAdd</PictureUploadPolicy>
+                           </UploadSiteHostedPicturesRequest>';
+
+        $boundary = "==Multipart_Boundary_x" . md5(mt_rand()) . "x";
+        $CRLF = "\r\n";
+
+        // The complete POST consists of an XML request plus the binary image separated by boundaries
+        $firstPart = '';
+        $firstPart .= "--" . $boundary . $CRLF;
+        $firstPart .= 'Content-Disposition: form-data; name="XML Payload"' . $CRLF;
+        $firstPart .= 'Content-Type: text/xml;charset=utf-8' . $CRLF . $CRLF;
+        $firstPart .= $xmlReq;
+        $firstPart .= $CRLF;
+
+        //$secondPart = '';
+        $secondPart = "--" . $boundary . $CRLF;
+        $secondPart .= 'Content-Disposition: form-data; name="dummy"; filename="dummy"' . $CRLF;
+        $secondPart .= "Content-Transfer-Encoding: binary" . $CRLF;
+        $secondPart .= "Content-Type: application/octet-stream" . $CRLF . $CRLF;
+        $secondPart .= $multiPartImageData;
+        $secondPart .= $CRLF;
+        $secondPart .= "--" . $boundary . "--" . $CRLF;
+
+        $request_body = $firstPart . $secondPart;
+
+        $respXmlStr = L5ebtapiController::multiPartRequest('UploadSiteHostedPictures', $request_body, $boundary);   // send multi-part request and get string XML response
+
+        if (stristr($respXmlStr, 'HTTP 404')) {
+
+            Log::error('eBay API Call: uploadSiteHostedPictures() 404 Not Found');
+
+            return 'Error: 404 Not Found. Please verify all eBay API settings are correct and try again.';
+
+        }
+
+        if ($respXmlStr == '') {
+
+            Log::error('eBay API Call: uploadSiteHostedPictures() Error sending request' .
+                'the XML response is an empty string');
+
+            return 'Error: The XML response is an empty string. Please verify all eBay API settings are correct and' .
+            'try again.';
+
+        }
+
+        $respXmlObj = simplexml_load_string($respXmlStr);
+
+        if($respXmlObj->SiteHostedPictureDetails->FullURL) {
+
+            $picURL = (string)$respXmlObj->SiteHostedPictureDetails->FullURL;
+
+        }
+        else {
+
+            $picURL = 'Error: An error occurred While processing the uploadSiteHostedPictures() request. Please' .
+                'verify all eBay API settings are correct and all input is correct and then try the request again.';
+
+        }
+
+        return $picURL;
+
+    }// END of uploadSiteHostedPictures($multiPartImageData, $image_name)
+
+    /**
+     * Method: request($call_name, $request_body) - Make an eBay API request.
+     *
+     * @param $call_name the eBay API call name
      * @param $request_body the body of the request
      * @return mixed
      */
-    public function request($method, $request_body)
+    public function request($call_name, $request_body)
     {
 
         $client = new Client();
@@ -230,7 +317,7 @@ class L5ebtapiController extends Controller
                     'X-EBAY-API-APP-NAME' => $this->api_app_id,
                     'X-EBAY-API-CERT-NAME' => $this->api_cert_id,
                     'X-EBAY-API-SITEID' => $this->api_site_id,
-                    'X-EBAY-API-CALL-NAME' => $method
+                    'X-EBAY-API-CALL-NAME' => $call_name
                 ),
                 'body' => $request_body
             ));
@@ -247,6 +334,50 @@ class L5ebtapiController extends Controller
 
         return $body;
 
-    }
+    }// END - request($call_name, $request_body)
 
-}
+    /**
+     * Method: multiPartRequest($call_name, $request_body, $boundary) - Make an eBay API multi-part request.
+     *
+     * @param $call_name the eBay API call name
+     * @param $request_body the body of the request
+     * @param $boundary the boundary for the multi-part data
+     * @return mixed
+     */
+    public function multiPartRequest($call_name, $request_body, $boundary)
+    {
+
+            $client = new Client();
+
+            try {
+
+                $response = $client->post($this->api_url, array(
+                    'verify' => $this->api_verify_ssl,
+                    'headers' => array(
+                        'HTTP' => '1.0',
+                        'Content-Type' => 'multipart/form-data; boundary=' . $boundary,
+                        'Content-Length' => strlen($request_body),
+                        'X-EBAY-API-COMPATIBILITY-LEVEL' => $this->api_compatibility_level,
+                        'X-EBAY-API-DEV-NAME' => $this->api_dev_id,
+                        'X-EBAY-API-APP-NAME' => $this->api_app_id,
+                        'X-EBAY-API-CERT-NAME' => $this->api_cert_id,
+                        'X-EBAY-API-SITEID' => $this->api_site_id,
+                        'X-EBAY-API-CALL-NAME' => $call_name
+                    ),
+                    'body' => $request_body
+                ));
+
+            } catch (\GuzzleHttp\Exception\ServerException $e) {
+
+                $response = $e->getResponse();
+
+                Log::warning($response->getBody()->getContents());
+
+            }
+
+            return $response->getBody()->getContents();
+
+    } // END - multiPartRequest($call_name, $request_body, $boundary)
+
+
+}// END of class L5ebtapiController