+++ CHANGELOG.md
@@ -7,13 +7,14 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## NEXT - 2016-07-18
 
 ### Added
-- eBay Trading API method getEbayOfficialTime().
+- Method: getEbayOfficialTime() - get the eBay official time API call.
+- Method: uploadSiteHostedPictures($multiPartImageData, $image_name) - Upload an image to the eBay Picture Service.
 
 ### Deprecated
 - Nothing
 
 ### Fixed
-- Cleaned up code.
+- Code documentation improved.
 
 ### Removed
 - Nothing